# openapi.yaml
openapi: 3.0.3
info:
  title: Taskryse API
  version: 1.0.0
  description: >
    Backend contract for Taskryse. Money in minor units (kobo/cents).
    Core rule: plans grant ELIGIBILITY only; earnings come from APPROVED work
    against FUNDED sponsor escrow; promo credits are non-withdrawable.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
paths:
...
  /assignments/{id}/autosave: { patch: { parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: { '200': { description: saved } } } }
  /assignments/{id}/submit: { post: { summary: Submit -> pending earnings (not withdrawable), parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: { '200': { description: under_review } } } }
...